+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `Image::on_load_layout` event fired after new image is loaded and layout.
 * Optimize image decoding, faster and less memory use.
     - Decodes directly to the IPC memory map.
     - Optimized EXIF orientation transforms.